+++ crates/cgp-field/src/traits/has_field_mut.rs
@@ -2,23 +2,22 @@ use core::marker::PhantomData;
 use core::ops::DerefMut;
 
 use crate::traits::has_field::HasField;
+use crate::FieldGetter;
 
 pub trait HasFieldMut<Tag>: HasField<Tag> {
-    fn get_field_mut(&mut self, tag: PhantomData<Tag>) -> &mut Self::Field;
+    fn get_field_mut(&mut self, tag: PhantomData<Tag>) -> &mut Self::Value;
 }
 
-pub trait MutFieldGetter<Context, Tag> {
-    type Field;
-
-    fn get_field_mut(context: &mut Context, tag: PhantomData<Tag>) -> &mut Self::Field;
+pub trait MutFieldGetter<Context, Tag>: FieldGetter<Context, Tag> {
+    fn get_field_mut(context: &mut Context, tag: PhantomData<Tag>) -> &mut Self::Value;
 }
 
-impl<Context, Tag, Target, Field> HasFieldMut<Tag> for Context
+impl<Context, Tag, Target, Value> HasFieldMut<Tag> for Context
 where
     Context: DerefMut<Target = Target>,
-    Target: HasFieldMut<Tag, Field = Field> + 'static,
+    Target: HasFieldMut<Tag, Value = Value> + 'static,
 {
-    fn get_field_mut(&mut self, tag: PhantomData<Tag>) -> &mut Self::Field {
+    fn get_field_mut(&mut self, tag: PhantomData<Tag>) -> &mut Self::Value {
         self.deref_mut().get_field_mut(tag)
     }
 }