+++ crates/cgp/src/lib.rs
@@ -1,3 +1,2 @@
-pub use cgp_core as core;
 pub use cgp_core::prelude;
-pub use cgp_extra as extra;
+pub use {cgp_core as core, cgp_extra as extra};