+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Pre-Release
 
+- Rename `HasField::Field` to `HasField::Value` - [#35](https://github.com/contextgeneric/cgp/pull/35)
+
 - Remove `Sized` constraint from `Async` trait - [#34](https://github.com/contextgeneric/cgp/pull/34)
 
 - Component pattern improvements - [#24](https://github.com/contextgeneric/cgp/pull/24)